+++ RJJ_ObjGen_MemManage.cpp
@@ -5,7 +5,7 @@ using namespace std;
 
 // functions using floats
 
-void InitObjGen(vector <object_props *> & detections, int & NOobj, int obj_limit, vector<int> & obj_ids, vector<int> & check_obj_ids, int * & data_metric, int * & xyz_order){
+void InitObjGen(vector <object_props *> & detections, int & NOobj, int obj_limit, vector<int> & obj_ids, vector<int> & check_obj_ids, size_t * & data_metric, int * & xyz_order){
 
   detections.reserve(1000);
   detections.resize(0);
@@ -16,12 +16,12 @@ void InitObjGen(vector <object_props *> & detections, int & NOobj, int obj_limit
   check_obj_ids.reserve(static_cast<int>(1.0E6));
   check_obj_ids.resize(0);
   NOobj = 0;
-  data_metric = new int[3];
+  data_metric = new size_t[3];
   xyz_order = new int[3];
   
 }
 
-void InitObjGen(vector <object_props *> & detections, long int & NOobj, int obj_limit, vector<long int> & obj_ids, vector<long int> & check_obj_ids, int * & data_metric, int * & xyz_order){
+void InitObjGen(vector <object_props *> & detections, long int & NOobj, int obj_limit, vector<long int> & obj_ids, vector<long int> & check_obj_ids, size_t * & data_metric, int * & xyz_order){
 
   detections.reserve(1000);
   detections.resize(0);
@@ -32,14 +32,14 @@ void InitObjGen(vector <object_props *> & detections, long int & NOobj, int obj_
   check_obj_ids.reserve(static_cast<int>(1.0E6));
   check_obj_ids.resize(0);
   NOobj = 0;
-  data_metric = new int[3];
+  data_metric = new size_t[3];
   xyz_order = new int[3];
   
 }
 
-void FreeObjGen(vector <object_props *> & detections, int * & data_metric, int * & xyz_order){
+void FreeObjGen(vector <object_props *> & detections, size_t * & data_metric, int * & xyz_order){
 
-  int i;
+  unsigned int i;
 
   for(i = 0; i < detections.size(); ++i){
 
@@ -53,7 +53,7 @@ void FreeObjGen(vector <object_props *> & detections, int * & data_metric, int *
 
 // functions using doubles
 
-void InitObjGen(vector <object_props_dbl *> & detections, int & NOobj, int obj_limit, vector<int> & obj_ids, vector<int> & check_obj_ids, int * & data_metric, int * & xyz_order){
+void InitObjGen(vector <object_props_dbl *> & detections, int & NOobj, int obj_limit, vector<int> & obj_ids, vector<int> & check_obj_ids, size_t * & data_metric, int * & xyz_order){
 
   detections.reserve(1000);
   detections.resize(0);
@@ -64,12 +64,12 @@ void InitObjGen(vector <object_props_dbl *> & detections, int & NOobj, int obj_l
   check_obj_ids.reserve(static_cast<int>(1.0E6));
   check_obj_ids.resize(0);
   NOobj = 0;
-  data_metric = new int[3];
+  data_metric = new size_t[3];
   xyz_order = new int[3];
   
 }
 
-void InitObjGen(vector <object_props_dbl *> & detections, long int & NOobj, int obj_limit, vector<long int> & obj_ids, vector<long int> & check_obj_ids, int * & data_metric, int * & xyz_order){
+void InitObjGen(vector <object_props_dbl *> & detections, long int & NOobj, int obj_limit, vector<long int> & obj_ids, vector<long int> & check_obj_ids, size_t * & data_metric, int * & xyz_order){
 
   detections.reserve(1000);
   detections.resize(0);
@@ -80,14 +80,14 @@ void InitObjGen(vector <object_props_dbl *> & detections, long int & NOobj, int 
   check_obj_ids.reserve(static_cast<int>(1.0E6));
   check_obj_ids.resize(0);
   NOobj = 0;
-  data_metric = new int[3];
+  data_metric = new size_t[3];
   xyz_order = new int[3];
   
 }
 
-void FreeObjGen(vector <object_props_dbl *> & detections, int * & data_metric, int * & xyz_order){
+void FreeObjGen(vector <object_props_dbl *> & detections, size_t * & data_metric, int * & xyz_order){
 
-  int i;
+  unsigned int i;
 
   for(i = 0; i < detections.size(); ++i){
